+++ migrations/20190506162321_Customer.js
@@ -2,7 +2,7 @@ exports.up = function(knex, Promise) {
     return knex.schema
     .createTable('Customer', function(table) {
         table.increments('id');
-        table.tinyInteger('gender').notNullable();
+        table.integer('gender').notNullable();
         table.string('name', 50).notNullable();
         table.string('email', 50).notNullable();
         table.text('address').notNullable();
